#14 replaced int? with int in services

+++ SoaWebsite.Web/Controllers/DevelopersController.cs
@@ -57,7 +57,7 @@ namespace SoaWebsite.Web.Controllers
             {
                 return NotFound();
             }
-            var added = service.AddSkill(idDeveloper, skill);
+            var added = service.AddSkill((int)idDeveloper, skill);
             if (ModelState.IsValid && added)
             {
                 return RedirectToAction("Index");
@@ -67,7 +67,7 @@ namespace SoaWebsite.Web.Controllers
 
         public IActionResult Details(int? idDeveloper)
         {
-            Developer developer = service.DeveloperWithSkillsById(idDeveloper);
+            Developer developer = service.DeveloperWithSkillsById((int)idDeveloper);
             if (developer == null)
             {
                 return NotFound();
@@ -77,7 +77,7 @@ namespace SoaWebsite.Web.Controllers
 
         public IActionResult Edit(int? idDeveloper)
         {
-            Developer developer = service.DeveloperWithSkillsById(idDeveloper);
+            Developer developer = service.DeveloperWithSkillsById((int)idDeveloper);
             if (developer == null)
             {
                 return NotFound();
@@ -118,7 +118,7 @@ namespace SoaWebsite.Web.Controllers
         // GET: Developers/Delete/5
         public IActionResult Delete(int? idDeveloper)
         {
-            var developer = service.DeveloperById(idDeveloper);
+            var developer = service.DeveloperById((int)idDeveloper);
             if (developer == null)
             {
                 return NotFound();
@@ -144,7 +144,7 @@ namespace SoaWebsite.Web.Controllers
         // GET: Skills/Delete/5
         public IActionResult DeleteSkill(int? idDeveloper, int? idSkill)
         {
-            var developerSkill = service.GetDeveloperSkill(idDeveloper, idSkill);
+            var developerSkill = service.GetDeveloperSkill((int)idDeveloper, (int)idSkill);
             if (developerSkill == null)
             {
                 return NotFound();
